Fixed Data Model Definitions.
Changed the order of the end point processing so that the correct data
model is produced and use subclasses to implement extensions to the base
data models. At this point the subclass mechanism depends on the data
models being generated in the correct order, where base classes are
generated before sub classes. This is done by processing the endpoints
in order: POST base endpoint, all item endpoints, GET base endpoint.
This seems to work for nfs exports, obviously won't work if the same
pattern that nfs exports uses is not repeated by the other endpoints.