Removed ResultsIdDirectoriesTotalUsage definition
The ResultsIdDirectoriesTotalUsage is not used by any end points, so I
removed it. I suspect it made it in there when a end point that fails
processing was partially processed (i.e. the schema was added to the
object definitions, but then the end point failed to fully finish
processing so it was never added to the "paths" section, so the data
model definition was orphaned).