Fix for PAPI schemas with no type or properties
An early adopter of the SDK found an issue with the /cluster/version and
/cluster/config related to the PAPI json description not being valid. So
I put in another hack to fix up the PAPI description so that it convert
correctly.

I also added in some code to better handle naming of objects that
consist of URLs where some part of the middle portion of the URL
consists of <*ID>. Instead of creating names like SnapshotsIdLock it
creates it as SnapshotsSnapshotLock.