Changed how API name is generated.
In order to simplify the function name (i.e. operationId) that gets
generated for really long end point URLs, I changed the way the API name
(i.e. the "tags" field) gets generated so that URLs that include a
middle parameter (i.e. a piece in the middle that is variable/dynamic)
get put into a separate API.