Fixes for issues found with latest Swagger Codegen
The latest version of the Swagger code generator complains about the
"pattern" parameters in our schemas because they don't use the "Perl"
style of specifying a regex. So I put in some code to fix this. I also
found, when trying to regenerate the 7.2 SDK, that I needed to have a
special list of excluded end points for 7.2 vs 8.0 because in version
7.2 the list of PAPI end points is sorted differently than in 8.0, so
the logic for picking the highest version end point does not work. So
now you have to use the excluded_end_points_7_2.json to exclude old end
points (luckily there aren't that many in 7.2).

Lastly, I added a configuration file for the Swagger code generator that
tells it to name the src directory for the Python code "isi_sdk" (that
way we don't have to do that change by hand).