Deals with required params better now
Fixed a couple of issues related to required parameters.

If the response contains a multi-type parameter (i.e. the type is a
list) whose type can be null and if the parameter is also required
then it causes Swagger to throw an exception. So now multi-type
parameters that are in response objects that are "required" by the PAPI
are not specified as required by the swagger config.

Also, now the full object model, including the required fields, are
taken into account when attempting to find a matching data model or a
data model to extend.