Updated the README.md and swagger codegen config
Updated the instructions for how to generate the Python bindings so that
it includes the use of the customer template in the swagger_templates
directory.

Also, wrote a couple of simple tests for the /cluster/config,
/cluster/version, and /quota end points to verify that my changes/fixes
to the swagger config generator actually fix those end points as
expected. Added test_event_eventgroup_ocurrences.py, but it doesn't
actually do anything useful yet because that end point is broken in 8.0.

Also, fixed the test_auth_groups.py, which was broken with the previous
update to the SDK.