Adds configure() and IsiApiClient_8_0 and IsiApiClient_7_2
The configure() function handles the details of picking best version of
the sdk to use for a particular cluster based on which version of the
sdk is installed and the version number of the cluster.

The IsiApiClient_8_0 and IsiApiClient_7_2 are subclasses of the Swagger
Codegen generated class ApiClient that allow the sdk to be used in a
multi-thread or multi-client scenario. Swagger Codegen's current
implementation of ApiClient uses a singleton for authentication
credential configuration, so its not possible to use it for
multi-thread scenarios where each thread wants to connect to a different
cluster. So these classes fix that issue.