Reverted templates to original versions.
I mistakenly used the templates from the latest version of Swagger
Codegen for my last commit, which ended up breaking things. So these are
the same changes on the templates to fix the /nfs/aliases/<AID> end
point, but applied to the version of the templates that we use in our
build (1939ce8e91f9de4a35ba3d90fd99a28f41c5035c).